Change balance type from decimal to float

+++ priv/repo/migrations/20220508003008_create_accounts_table.exs
@@ -3,7 +3,7 @@ defmodule Argentaria.Repo.Migrations.CreateAccountsTable do
 
   def change do
     create table(:accounts) do
-      add(:balance, :decimal)
+      add(:balance, :float)
       timestamps()
     end
   end
